requirements.txt
fastapi[all]
Pillow
numpy
pydantic